requirements.txt
gunicorn
ffmpeg-python
google-genai
openai
datetime
flask
flask-cors
python-dotenv
pydantic
uuid
asyncio
requests
elevenlabs
ffmpeg-python
Pillow